Move sortition code to pkg/sortition
This is a common package that will be used by beacon and ecdsa modules.
