+++ pkg/chain/random-beacon/chain.go
@@ -1,8 +1,6 @@
 package beacon
 
-import (
-	"github.com/keep-network/keep-core/pkg/chain/sortition"
-)
+import "github.com/keep-network/keep-core/pkg/sortition"
 
 // Handle for interaction with the Random Beacon module contracts.
 type Handle interface {
